[MIG] auth_keycloak: Migration to 16.0
