--- auth_keycloak/examples/odoo-client.json
@@ -1,48 +1,0 @@
-{
-  "clientId": "odoo",
-  "baseUrl": "/auth/realms/master/odoo",
-  "surrogateAuthRequired": false,
-  "enabled": true,
-  "clientAuthenticatorType": "client-secret",
-  "redirectUris": [
-    "http://localhost:8069/*",
-    "http://localhost:8501/*",
-    "http://localhost/*"
-  ],
-  "webOrigins": [],
-  "notBefore": 0,
-  "bearerOnly": false,
-  "consentRequired": false,
-  "standardFlowEnabled": true,
-  "implicitFlowEnabled": true,
-  "directAccessGrantsEnabled": true,
-  "serviceAccountsEnabled": false,
-  "publicClient": false,
-  "frontchannelLogout": false,
-  "protocol": "openid-connect",
-  "attributes": {
-    "saml.assertion.signature": "false",
-    "saml.force.post.binding": "false",
-    "saml.multivalued.roles": "false",
-    "saml.encrypt": "false",
-    "saml.server.signature": "false",
-    "saml.server.signature.keyinfo.ext": "false",
-    "exclude.session.state.from.auth.response": "true",
-    "saml_force_name_id_format": "false",
-    "saml.client.signature": "false",
-    "tls.client.certificate.bound.access.tokens": "false",
-    "saml.authnstatement": "false",
-    "display.on.consent.screen": "false",
-    "saml.onetimeuse.condition": "false"
-  },
-  "authenticationFlowBindingOverrides": {},
-  "fullScopeAllowed": true,
-  "nodeReRegistrationTimeout": -1,
-  "defaultClientScopes": ["role_list", "profile", "email"],
-  "optionalClientScopes": ["address", "phone", "offline_access"],
-  "access": {
-    "view": true,
-    "configure": true,
-    "manage": true
-  }
-}
