--- auth_keycloak/examples/requirements.txt
@@ -1,2 +1,0 @@
-requests
-click
