--- auth_keycloak/exceptions.py
@@ -1,2 +1,0 @@
-class OAuthError(Exception):
-    pass
