+++ auth_keycloak/readme/HISTORY.rst
@@ -1,3 +1,8 @@
+16.0.1.0.0 2026-02-20
+~~~~~~~~~~~~~~~~~~~~~
+
+* v16 migration only implementing user creation, everything else is done by auth_oidc
+
 10.0.1.0.0 2018-10-17
 ~~~~~~~~~~~~~~~~~~~~~
