+++ auth_keycloak/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+auth_keycloak.access_auth_keycloak_sync_wiz,access_auth_keycloak_sync_wiz,model_auth_keycloak_sync_wiz,base.group_system,1,1,1,1
+auth_keycloak.access_auth_keycloak_create_wiz,access_auth_keycloak_create_wiz,model_auth_keycloak_create_wiz,base.group_system,1,1,1,1
