+++ auth_keycloak/tests/__init__.py
@@ -1,3 +1,2 @@
-from . import test_auth
 from . import test_wizard_sync
 from . import test_wizard_create
