+++ auth_keycloak/views/auth_oauth_views.xml
@@ -5,18 +5,20 @@
     <field name="model">auth.oauth.provider</field>
     <field name="inherit_id" ref="auth_oauth.view_oauth_provider_form" />
     <field name="arch" type="xml">
-      <field name="client_id" position="after">
-        <field name="client_secret" />
-      </field>
       <xpath expr="//field[@name='auth_endpoint']/ancestor::group[1]" position="after">
         <field name="id" invisible="1" />
         <!-- TODO: any good way to hide/show this properly -->
         <group string="Users management (Keycloak)">
-          <field name="users_endpoint" />
+          <field
+                        name="users_endpoint"
+                        placeholder="https://$keycloak_host/admin/realms/$realm/users"
+                    />
           <field name="superuser" />
-          <field name="superuser_pwd" />
+          <field name="superuser_pwd" password="True" />
           <field name="users_management_enabled" />
           <button
+                        colspan="2"
+                        class="btn btn-primary"
                         string="Sync users"
                         context="{'default_provider_id': id}"
                         name="%(auth_keycloak.keycloak_sync_users)d"
