+++ auth_keycloak/views/res_users_views.xml
@@ -7,7 +7,6 @@
     <field name="inherit_id" ref="base.view_users_form" />
     <field name="arch" type="xml">
       <header position="inside">
-        <field name="oauth_uid" invisible="1" readonly="1" />
         <button
                     string="Push to Keycloak"
                     type="object"
