+++ auth_keycloak/wizard/__init__.py
@@ -1,1 +1,3 @@
-from . import keycloak_sync_wiz
+from . import auth_keycloak_sync_mixin
+from . import auth_keycloak_create_wiz
+from . import auth_keycloak_sync_wiz
