+++ auth_keycloak/wizard/auth_keycloak_create_wiz.xml
@@ -11,6 +11,7 @@
             <group name="provider">
               <field name="provider_id" />
               <div
+                                colspan="2"
                                 class="alert alert-warning"
                                 role="alert"
                                 attrs="{'invisible': [('management_enabled', '=', True)]}"
@@ -18,6 +19,7 @@
                 <strong>Users managerment not enabled!</strong>
                 You must configure "Users management" parameters on selected provider.
               </div>
+              <field name="login_match_key" />
             </group>
             <group
                             name="metadata"
@@ -25,12 +27,11 @@
                         >
               <field name="endpoint" />
               <field name="user" />
-              <field name="pwd" />
-              <field name="login_match_key" />
+              <field name="pwd" password="True" />
             </group>
           </group>
           <group name="users">
-            <field name="user_ids" readonly="1">
+            <field name="user_ids" readonly="1" nolabel="1" colspan="2">
               <tree>
                 <field name="name" />
                 <field name="login" />
@@ -41,12 +42,13 @@
           </group>
           <footer>
             <button
+                            class="btn btn-primary"
                             name="button_create_user"
                             string="Create"
                             attrs="{'invisible': [('management_enabled', '=', False)]}"
                             type="object"
                         />
-            <button string="Cancel" class="btn-default" special="cancel" />
+            <button string="Cancel" class="btn btn-secondary" special="cancel" />
           </footer>
         </sheet>
       </form>
@@ -56,7 +58,6 @@
   <record id="keycloak_create_users" model="ir.actions.act_window">
     <field name="name">Push users to Keycloak</field>
     <field name="res_model">auth.keycloak.create.wiz</field>
-    <field name="view_type">form</field>
     <field name="view_mode">tree,form</field>
     <field name="view_id" ref="auth_keycloak_create_wiz" />
     <field name="target">new</field>
